+++ src/components/SidePanel/FilePathLink.tsx
@@ -5,9 +5,8 @@
  */
 
 import React from 'react';
-import Button from 'react-bootstrap/Button';
 import FormLabel from 'react-bootstrap/FormLabel';
-import { truncateMiddle } from 'pc-nrfconnect-shared';
+import { Button, truncateMiddle } from 'pc-nrfconnect-shared';
 
 import { openInFolder } from '../../utils/fileUtils';